# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/sucursales/{id}": {
      "get": {
        "tags": [
          "sucursal-controller"
        ],
        "summary": "Obtener una sucursal por su ID",
        "description": "Retorna una sucursal específica según su ID.",
        "operationId": "obtenerPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Sucursal no encontrada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Sucursal"
                }
              }
            }
          },
          "200": {
            "description": "Sucursal encontrada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Sucursal"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "sucursal-controller"
        ],
        "summary": "Actualizar una sucursal existente",
        "description": "Actualiza una sucursal existente según su ID.",
        "operationId": "actualizarSucursal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sucursal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Sucursal no encontrada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Sucursal"
                }
              }
            }
          },
          "200": {
            "description": "Sucursal actualizada exitosamente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Sucursal"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sucursal-controller"
        ],
        "summary": "Eliminar una sucursal",
        "description": "Elimina una sucursal según su ID.",
        "operationId": "eliminarSucursal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Sucursal no encontrada."
          },
          "200": {
            "description": "Sucursal eliminada exitosamente."
          }
        }
      }
    },
    "/api/reservas/{id}": {
      "get": {
        "tags": [
          "reserva-controller"
        ],
        "operationId": "obtenerPorId_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Reserva"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "reserva-controller"
        ],
        "operationId": "actualizarReserva",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reserva"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Reserva"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reserva-controller"
        ],
        "operationId": "eliminarReserva",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/proveedores/{id}": {
      "get": {
        "tags": [
          "proveedor-controller"
        ],
        "operationId": "obtenerPorId_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Proveedor"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "proveedor-controller"
        ],
        "operationId": "actualizarProveedor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proveedor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Proveedor"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "proveedor-controller"
        ],
        "operationId": "eliminarProveedor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/productos/{id}": {
      "get": {
        "tags": [
          "producto-controller"
        ],
        "summary": "Obtener un producto por su ID",
        "description": "Retorna un producto específico según su ID con enlaces HATEOAS.",
        "operationId": "obtenerPorId_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProducto"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProducto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "producto-controller"
        ],
        "summary": "Actualizar un producto existente",
        "description": "Actualiza un producto existente según su ID y lo retorna con enlaces HATEOAS.",
        "operationId": "actualizarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProducto"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProducto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "producto-controller"
        ],
        "summary": "Eliminar un producto",
        "description": "Elimina un producto según su ID.",
        "operationId": "eliminarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado exitosamente."
          },
          "404": {
            "description": "Producto no encontrado."
          }
        }
      }
    },
    "/api/pedidos/{id}": {
      "get": {
        "tags": [
          "pedido-controller"
        ],
        "summary": "Obtener un pedido por su ID",
        "description": "Retorna un pedido específico según su ID con enlaces HATEOAS.",
        "operationId": "obtenerPorId_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Pedido no encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          },
          "200": {
            "description": "Pedido encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "pedido-controller"
        ],
        "summary": "Actualizar un pedido existente",
        "description": "Actualiza un pedido existente según su ID y lo retorna con enlaces HATEOAS.",
        "operationId": "actualizarPedido",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pedido"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Pedido actualizado exitosamente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pedido-controller"
        ],
        "summary": "Eliminar un pedido",
        "description": "Elimina un pedido según su ID.",
        "operationId": "eliminarPedido",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Pedido no encontrado."
          },
          "204": {
            "description": "Pedido eliminado exitosamente."
          }
        }
      }
    },
    "/api/inventarios/{id}": {
      "get": {
        "tags": [
          "inventario-controller"
        ],
        "operationId": "obtenerPorId_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "inventario-controller"
        ],
        "operationId": "actualizarInventario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "inventario-controller"
        ],
        "operationId": "eliminarInventario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clientes/{id}": {
      "get": {
        "tags": [
          "cliente-controller"
        ],
        "summary": "Obtener un cliente por su ID",
        "description": "Retorna un cliente específico según su ID con enlaces HATEOAS.",
        "operationId": "obtenerPorId_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Cliente no encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelCliente"
                }
              }
            }
          },
          "200": {
            "description": "Cliente encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelCliente"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cliente-controller"
        ],
        "summary": "Actualizar un cliente existente",
        "description": "Actualiza un cliente existente según su ID y lo retorna con enlaces HATEOAS.",
        "operationId": "actualizarCliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cliente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Cliente no encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelCliente"
                }
              }
            }
          },
          "200": {
            "description": "Cliente actualizado exitosamente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelCliente"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cliente-controller"
        ],
        "summary": "Eliminar un cliente",
        "description": "Elimina un cliente según su ID.",
        "operationId": "eliminarCliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente eliminado exitosamente."
          },
          "404": {
            "description": "Cliente no encontrado."
          }
        }
      }
    },
    "/api/categorias/{id}": {
      "get": {
        "tags": [
          "categoria-controller"
        ],
        "operationId": "obtenerPorId_8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "categoria-controller"
        ],
        "operationId": "actualizarCategoria",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categoria"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Categoria"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "categoria-controller"
        ],
        "operationId": "eliminarCategoria",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/productos": {
      "get": {
        "tags": [
          "producto-controller"
        ],
        "summary": "Obtener todos los productos",
        "description": "Retorna una lista de todos los productos con enlaces HATEOAS.",
        "operationId": "obtenerTodos_2",
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelProducto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "producto-controller"
        ],
        "summary": "Crear un nuevo producto",
        "description": "Crea un nuevo producto y lo retorna con enlaces HATEOAS.",
        "operationId": "crearProducto",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto creado exitosamente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProducto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notificaciones/{id}": {
      "get": {
        "tags": [
          "notificacion-controller"
        ],
        "operationId": "obtenerPorId_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Notificacion"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notificacion-controller"
        ],
        "operationId": "eliminarNotificacion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/inventarios/producto/{productoId}": {
      "get": {
        "tags": [
          "inventario-controller"
        ],
        "operationId": "obtenerPorProductoId",
        "parameters": [
          {
            "name": "productoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Inventario"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Sucursal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        }
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
...
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "precio": {
            "type": "number"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          }
        }
      },
      "Proveedor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "contacto": {
            "type": "string"
          }
        }
      },
      "Reserva": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fechaReserva": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          }
        }
      },
...
      "EntityModelProducto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "precio": {
            "type": "number"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
      "Pedido": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "listaProductos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Producto"
            }
          }
        }
      },
      "EntityModelPedido": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "listaProductos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Producto"
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Inventario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "umbral": {
            "type": "integer",
            "format": "int32"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "sucursal": {
            "$ref": "#/components/schemas/Sucursal"
          }
        }
      },
...
    }
  }
}